Fix our broken code
So the issue for those curious was javascript scoping :/. Turns out that
variables are scoped globaly so unless you play to shadow other variables
you can esily mess your self up by using the same variable names with out
care. In this particular case the t we where using and updating in
findInverse was actaully  the global t which means that the t we needed
for our algorithm was getting mercelesly trampeled for exact line fix see
line number 119. Also updated some comments so that they are nicer for
debugging. next step extract this code into another file and make a
library for our encryption mechanisim.